Add files via upload
This code implements three options. The firsts one shows how to use a groupbox with checkboxes. The second one shows how to use graphics with matplotlib. And the third one shows how to implement a graphic with parameters. This last option implements all the previous concepts describe in the tutorial.